video: tegra: host: Fix compile warnings
Some "unsigned long"s were converted to u32. for_each_set_bit()
requires unsigned long, so convert some instanced back.

Fix how include files are included.

Forward declare nvhost_set_error_notifier.

Fix parameter to nvhost_memmgr_put() to be handle and not id.

Fix tracing to use %llx and casting to u64 to print dma_addr_t.

Reviewed-on: http://git-master/r/301743

+++ drivers/video/tegra/host/nvhost_intr.c
@@ -281,7 +281,7 @@ irqreturn_t nvhost_syncpt_thresh_fn(void *dev_id)
 irqreturn_t nvhost_intr_irq_fn(int irq, void *dev_id)
 {
 	struct nvhost_intr *intr = dev_id;
-	u32 intstat = intr->intstatus;
+	unsigned long intstat = intr->intstatus;
 	int i;
 
 	for_each_set_bit(i, &intstat, 32) {
